feat: Update axios dependency to version 1.7.2 and refactor executedSeed method. And initialice the axios in the constructor of the service
This commit updates the axios dependency in the package.json file to version 1.7.2. It also refactors the executedSeed method in the seed.service.ts file to use async/await and axios for fetching data from an external API. These changes improve the code by using the latest version of axios and modernizing the asynchronous data fetching approach.  Also